Split out visualization code in preparation for extraction
